move CategoryDetailsUiState to screen package
