+++ app/src/main/java/com/example/tudeeapp/presentation/screen/categoryDetails/CategoryDetailsUiState.kt
@@ -1,4 +1,4 @@
-package com.example.tudeeapp.presentation.screen.categoryDetails.state
+package com.example.tudeeapp.presentation.screen.categoryDetails
 
 import com.example.tudeeapp.domain.models.Category
 import com.example.tudeeapp.domain.models.Task
